Move commit parser under scan_worker
